+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- display undeliverable message instead of a alert
+
 ## [0.13.6] - 2023-09-12
 
 ### Fixed